refactor: renamed constant file
