+++ react-client/src/routes/my-pets/petId/utils/petServiceDetailsConstants.tsx
